# docs/swagger.yaml
basePath: /api
definitions:
  rest.AnyToFiatCourseSlice:
    description: has additional fields consist of currency code and amount, but cant
      image here
    properties:
      base_currency:
        type: string
      date:
        type: string
    type: object
  rest.AnyToFiatHistoryPage:
    properties:
      history:
        items:
          $ref: '#/definitions/rest.AnyToFiatCourseSlice'
        type: array
      total:
        type: integer
    type: object
  rest.AnyToFiatPaginationRequest:
    properties:
      from_date:
        type: string
      limit:
        type: integer
      offset:
        type: integer
    type: object
  rest.BtcUsdtCourseSlice:
    properties:
      timestamp:
        type: integer
      value:
        type: number
    type: object
  rest.BtcUsdtHistoryPage:
    properties:
      history:
        items:
          $ref: '#/definitions/rest.BtcUsdtCourseSlice'
        type: array
      total:
        type: integer
    type: object
  rest.BtcUsdtPaginationRequest:
    properties:
      from_timestamp:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
    type: object
  rest.UnexpectedResult:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
  title: Currencies service api
  version: "1.0"
paths:
  /btcusdt:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.BtcUsdtCourseSlice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
      summary: get last BTC USDT course
    post:
      consumes:
      - application/json
      parameters:
      - description: pagination
        in: body
        name: pagination
        required: true
        schema:
          $ref: '#/definitions/rest.BtcUsdtPaginationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.BtcUsdtHistoryPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
      summary: get BTC USDT course history with pagination
  /currencies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AnyToFiatCourseSlice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
      summary: get last RUB to fiat course
    post:
      consumes:
      - application/json
      parameters:
      - description: pagination
        in: body
        name: pagination
        required: true
        schema:
          $ref: '#/definitions/rest.AnyToFiatPaginationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AnyToFiatHistoryPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
      summary: get RUB to fiat course history with pagination
  /latest:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AnyToFiatCourseSlice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.UnexpectedResult'
      summary: get last fiat to BTC course
schemes:
- http
swagger: "2.0"